Extract DynamicProperty into its own file
This makes it easy to support SwiftPM (for excluding the file which need the Objective-C implementations).
